requirements.txt
google-genai
black
isort
ruff
fastapi
uvicorn
pydantic
aiohttp
nltk
pypdf
python-dotenv
uvloop
sentence-transformers
pinecone
